renamed Game.h -> Utils.h. Updated the readme. The readme should be broken up into a book or something at this rate. I compeltely glossed over RenderEntries and how the renderer works now.
