+++ ExampleDemo/Scenes/AboutScene.h
@@ -10,7 +10,7 @@
 #include <Segues/BlendFadeIn.h>
 #include <Segues/Cube3D.h>
 #include <Swoosh/ActivityController.h>
-#include <Swoosh/Game.h>
+#include <Swoosh/Utils.h>
 #include <SFML/Graphics.hpp>
 #include <SFML/Audio.hpp>
 #include <iostream>
