+++ ExampleDemo/Scenes/GameplayScene.h
@@ -8,7 +8,7 @@
 
 #include <Segues/Checkerboard.h>
 #include <Swoosh/ActivityController.h>
-#include <Swoosh/Game.h>
+#include <Swoosh/Utils.h>
 #include <SFML/Graphics.hpp>
 #include <SFML/Audio.hpp>
 #include <iostream>
