+++ ExampleDemo/Scenes/HiscoreScene.h
@@ -6,7 +6,7 @@
 #include "../SaveFile.h"
 
 #include <Swoosh/ActivityController.h>
-#include <Swoosh/Game.h>
+#include <Swoosh/Utils.h>
 #include <SFML/Graphics.hpp>
 #include <SFML/Audio.hpp>
 #include <Segues/SlideIn.h>
