+++ ExampleDemo/Scenes/MainMenuScene.h
@@ -41,7 +41,7 @@
 // end segue effects
 
 #include <Swoosh/ActivityController.h>
-#include <Swoosh/Game.h>
+#include <Swoosh/Utils.h>
 #include <SFML/Graphics.hpp>
 #include <SFML/Audio.hpp>
 #include <iostream>
