+++ src/Segues/HorizontalSlice.h
@@ -1,6 +1,6 @@
 #pragma once
 #include <Swoosh/Segue.h>
-#include <Swoosh/Game.h>
+#include <Swoosh/Utils.h>
 #include <Swoosh/Ease.h>
 
 using namespace sw;
