+++ src/Swoosh/Utils.h
